control commit

+++ app/src/main/java/com/caparepa/companionfou/data/db/entity/basic/BasicCommandCode.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "basic_command_code")
 data class BasicCommandCode(
     val id: Long? = null,
-    val collectionNo: Int? = null,
+    val collectionNo: Long? = null,
     val name: String? = null,
     val rarity: Int? = null,
     val face: String? = null
