+++ app/src/main/java/com/caparepa/companionfou/data/model/nice/craftessence/CraftEssenceItem.kt
@@ -7,7 +7,7 @@ data class CraftEssenceItem(
     @SerializedName("id")
     val id: Long? = null,
     @SerializedName("collectionNo")
-    val collectionNo: Int? = null,
+    val collectionNo: Long? = null,
     @SerializedName("name")
     val name: String? = null,
     @SerializedName("type")
