+++ app/src/main/java/com/caparepa/companionfou/data/model/nice/servant/ServantItem.kt
@@ -6,7 +6,7 @@ data class ServantItem(
     @SerializedName("id")
     val id: Long? = null,
     @SerializedName("collectionNo")
-    val collectionNo: Int? = null,
+    val collectionNo: Long? = null,
     @SerializedName("name")
     val name: String? = null,
     @SerializedName("className")
